#2231 add TestTable
